backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
python-decouple==3.8

# File Storage (AWS S3)
boto3==1.34.34
django-storages==1.14.2

# Image Processing
Pillow==10.4.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
django-extensions==3.2.3
ipython==8.21.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0
